more stride: u32

+++ src/pixel.rs
@@ -109,8 +109,8 @@ pub trait Pixel:
     /// Converts stride in pixels to stride in bytes.
     #[inline]
     #[allow(clippy::wrong_self_convention)]
-    fn to_asm_stride(in_stride: usize) -> isize {
-        (in_stride * size_of::<Self>()) as isize
+    fn to_asm_stride(in_stride: u32) -> isize {
+        (in_stride * size_of::<Self>() as u32) as isize
     }
 }
